+++ packages/kitchensink/service-apollo/index.js
@@ -1,4 +1,4 @@
-const forrestjs = require('@forrestjs/hooks');
+const forrestjs = require('@forrestjs/core');
 const serviceFastify = require('@forrestjs/service-fastify');
 const serviceApollo = require('@forrestjs/service-apollo');
 const gql = require('graphql-tag');