+++ packages/kitchensink/service-fastify-gql/index.js
@@ -1,4 +1,4 @@
-const forrestjs = require('@forrestjs/hooks');
+const forrestjs = require('@forrestjs/core');
 const serviceFastify = require('@forrestjs/service-fastify');
 const serviceFastifyGQL = require('@forrestjs/service-fastify-gql');