+++ packages/kitchensink/service-fastify-healthz/index.js
@@ -1,4 +1,4 @@
-const forrestjs = require('@forrestjs/hooks');
+const forrestjs = require('@forrestjs/core');
 const serviceLogger = require('@forrestjs/service-logger');
 const serviceFastify = require('@forrestjs/service-fastify');
 const serviceFastifyHealthz = require('@forrestjs/service-fastify-healthz');