+++ packages/kitchensink/service-jwt/index.js
@@ -1,4 +1,4 @@
-const forrestjs = require('@forrestjs/hooks');
+const forrestjs = require('@forrestjs/core');
 const serviceJwt = require('@forrestjs/service-jwt');
 const serviceFastify = require('@forrestjs/service-fastify');