+++ packages/kitchensink/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v4.0.0
 
-- Added examples for `@forrestjs/hooks`
+- Added examples for `@forrestjs/core`
 - Added examples for `@forrestjs/service-fastify`
 - Added examples for `@forrestjs/service-jwt`
 - Added examples for `@forrestjs/service-hash`